requirements.txt
aiofiles==24.1.0
aiogram==3.14.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
configobj==5.0.9
configparser==7.1.0
etelemetry==0.3.1
filelock==3.16.1
fitz==0.0.1.dev2
frozenlist==1.5.0
httplib2==0.22.0
idna==3.10
importlib_resources==6.4.5
isodate==0.6.1
looseversion==1.3.0
lxml==5.3.0
magic-filter==1.0.12
multidict==6.1.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.0
numpy==1.26.4
packaging==24.1
pandas==2.2.3
pathlib==1.0.1
pillow==11.0.0
propcache==0.2.0
prov==2.0.1
puremagic==1.28
pydantic==2.9.2
pydantic_core==2.23.4
pydot==3.0.2
PyMuPDF==1.24.13
pyparsing==3.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyxnat==1.6.2
rdflib==6.3.2
requests==2.32.3
scipy==1.14.1
simplejson==3.19.3
six==1.16.0
traits==6.4.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
yarl==1.17.1
